Moved files to correct position
Added a new "functionality" directory and moved the relevant files there
