+++ requirements.txt
@@ -1,2 +1,3 @@
 Django==5.1.2
 python-dotenv~=1.0.1
+requests~=2.32.3
